Added words on linking feature importance, increased fontsize on violin plots
Is this addition ok?
"It is important to note the difference here between classes with good feature discrimination, as observed in the violin plots, and feature importance for the classifier, as estimated above. The violin plots demonstrate the relative biophysical importance of each feature to class types. However, the classifier estimates of feature importance are unrelated to biophysical processes. Instead, they represent the relative importance of each feature to the nominated classifier and may have no rational explanation or connection to physical processes - such is the mystery of machine learning models."