Scalable ML on the ODC (#773)
* start porting scalable-ml

* predict working

* all notebooks working now

* remove hash from classification.py

* convert to running on WA only

* some typos

* fixed plot and title order

* Added words on linking feature importance, increased fontsize on violin plots

Is this addition ok?
"It is important to note the difference here between classes with good feature discrimination, as observed in the violin plots, and feature importance for the classifier, as estimated above. The violin plots demonstrate the relative biophysical importance of each feature to class types. However, the classifier estimates of feature importance are unrelated to biophysical processes. Instead, they represent the relative importance of each feature to the nominated classifier and may have no rational explanation or connection to physical processes - such is the mystery of machine learning models."

* minor editing to markdown in Plotting Results cell

* tidy markdown

* Consistency updates, add notebooks to readthedocs index

* Restore results files

Co-authored-by: Sean Chua <sean.chua@ga.gov.au>
Co-authored-by: Claire Phillips <claire.phillips@ga.gov.au>
Co-authored-by: robbibt <Robbi.BishopTaylor@ga.gov.au>