updates to landcover notebooks and code for new Collection 3
The changes affect the notebooks that use the landcover product. These have been updated and now they use the new collection 3 landcover. 
The python file containing the functions used in these notebooks (landcover.py) has also been updated. The script and its functions are stable and has also been tested for functionalities that are not currently in the notebooks but could be explored if users decide to expand from what presented in the notebooks. However, it could still be tidied up a bit (we are working on it on a separate branch)


---------

Co-authored-by: Jenna Guffogg <jenna.guffogg@gmail.com>